+++ example/lib/main.dart
@@ -45,10 +45,7 @@ class _HomePageState extends State<HomePage> {
         EasyTableColumn(
             name: 'Gender',
             width: 80,
-            cellBuilder: (context, row) => EasyTableCell(
-                child: row.male
-                    ? const Icon(Icons.male)
-                    : const Icon(Icons.female))),
+            iconValue: (row) => row.male ? Icons.male : Icons.female),
         EasyTableColumn(
             name: 'Race', width: 100, stringValue: (row) => row.race),
         EasyTableColumn(