+++ lib/src/value_mapper.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 /// Signature for a function that maps an [int] value of a row.
 ///
 /// Used by [EasyTableColumn].
@@ -17,3 +19,8 @@ typedef EasyTableStringValueMapper<ROW> = String? Function(ROW row);
 ///
 /// Used by [EasyTableColumn].
 typedef EasyTableObjectValueMapper<ROW> = Object? Function(ROW row);
+
+/// Signature for a function that maps a [IconData] value of a row.
+///
+/// Used by [EasyTableColumn].
+typedef EasyTableIconValueMapper<ROW> = IconData? Function(ROW row);